requirements.txt
django
djangorestframework
flake8
python-memcached








